+++ baselib/src/channelinfo.h
@@ -53,7 +53,7 @@ class BASELIB_EXPORT ChannelInfo
         const QString & talkingto_id() const;
         const QString & direction() const;
         const QString & commstatus() const;
-        int timestamp() const;
+        double timestamp() const;
 
         const QString peerdisplay() const;
         int linenumber() const;
@@ -74,7 +74,7 @@ class BASELIB_EXPORT ChannelInfo
         QString m_commstatus;
         QString m_talkingto_kind;
         QString m_talkingto_id;
-        int m_timestamp;
+        double m_timestamp;
 
         int m_linenumber;
         bool m_ismonitored;
